requirements.txt
absl-py==0.15.0
aiohttp==3.8.0
aiosignal==1.2.0
asgiref==3.4.1
astunparse==1.6.3
async-timeout==4.0.0
attrs==21.2.0
backports.entry-points-selectable==1.1.1
cachetools==4.2.4
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
configparser==5.1.0
cycler==0.11.0
datasets==1.15.1
dill==0.3.4
distlib==0.3.4
Django==3.2.8
django-cors-headers==3.10.0
django-environ==0.4.5
djangorestframework==3.12.4
docker-pycreds==0.4.0
environ==1.0
et-xmlfile==1.1.0
filelock==3.3.2
Flask==2.0.2
flatbuffers==2.0
fonttools==4.28.5
frozenlist==1.2.0
fsspec==2021.11.0
future==0.18.2
gast==0.4.0
gcloud==0.17.0
gensim==4.1.2
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.41.1
h5py==3.6.0
httplib2==0.20.2
huggingface-hub==0.1.2
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
jws==0.1.3
keras==2.7.0
Keras-Preprocessing==1.1.2
keytotext==1.5.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.5.1
multidict==5.2.0
multiprocess==0.70.12.2
nltk==3.6.5
numpy==1.21.3
oauth2client==3.0.0
oauthlib==3.1.1
openai==0.11.4
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.2
pandas==1.3.4
pandas-stubs==1.2.0.39
pathtools==0.1.2
Pillow==8.4.0
platformdirs==2.4.0
promise==2.3
protobuf==3.19.1
psutil==5.8.0
pyarrow==6.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
pyDeprecate==0.3.1
pyparsing==2.4.7
Pyrebase==3.0.27
python-dateutil==2.8.2
python-jwt==2.0.1
pytorch-lightning==1.5.1
pytz==2020.5
PyYAML==6.0
razorpay==1.2.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
rsa==4.7.2
sacremoses==0.0.46
scipy==1.7.3
sentencepiece==0.1.96
sentry-sdk==1.4.3
shortuuid==1.0.7
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
sqlparse==0.4.1
stripe==2.62.0
subprocess32==3.5.4
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
tokenizers==0.10.3
torch==1.10.0
torchmetrics==0.6.0
tqdm==4.62.3
transformers==4.12.3
typing-extensions==3.10.0.2
urllib3==1.26.2
virtualenv==20.10.0
wandb==0.12.6
Werkzeug==2.0.2
wrapt==1.13.3
xxhash==2.0.2
yarl==1.7.2
yaspin==2.1.0
